hugo: update from 0.57.2 to 0.59.1
That's the last version before switching from Blackfriday to Goldmark in
v0.60.0.

https://github.com/gohugoio/hugo/releases/tag/v0.60.0

It happens that my websites are not compatible with Goldmark, so I
wanted to do a last hugo-tidy push with the last compatible Hugo
version, so that I can then do a switch over with less things changing
underneath.